+++ playground/nextjs-app-router/components/DemoOptions.tsx
@@ -1,7 +1,6 @@
 import { ComponentMode } from '@/components/form/component-mode';
 import { ComponentTheme } from '@/components/form/component-theme';
 import { PaymasterUrl } from '@/components/form/paymaster';
-import { OnchainKitComponent } from './AppProvider';
 import { ActiveComponent } from './form/active-component';
 import { Chain } from './form/chain';
 import { CheckoutOptions } from './form/checkout-options';
@@ -10,6 +9,7 @@ import { NFTOptions } from './form/nft-options';
 import { SwapConfig } from './form/swap-config';
 import { TransactionOptions } from './form/transaction-options';
 import { WalletType } from './form/wallet-type';
+import { OnchainKitComponent } from '@/types/onchainkit';
 
 const COMMON_OPTIONS = [
   ActiveComponent,
